Fixed fog and made update independent of render
While in theory and in practice, making update independent of render makes it faster, it actually slows visual performance (without even slowing FPS or UPS). This has to be investigated

+++ cubyz-client/assets/cubyz/shaders/default/fragment.fs
@@ -52,7 +52,7 @@ struct Material
 struct Fog {
 	int active;
 	vec3 colour;
-	int density;
+	float density;
 };
 
 uniform sampler2D texture_sampler;
